+++ packages/app-admin-cognito/src/views/RequireNewPassword.tsx
@@ -6,7 +6,6 @@ import { Input } from "@webiny/ui/Input";
 import { Grid, Cell } from "@webiny/ui/Grid";
 import { Typography } from "@webiny/ui/Typography";
 import { Elevation } from "@webiny/ui/Elevation";
-import { useAuthenticator } from "@webiny/app-cognito-authenticator/hooks/useAuthenticator";
 import { useRequireNewPassword } from "@webiny/app-cognito-authenticator/hooks/useRequireNewPassword";
 import StateContainer from "./StateContainer";
 import { alignRight, InnerContent, Title } from "./StyledComponents";
@@ -17,10 +16,9 @@ const sentenceCase = (str: string) => {
 };
 
 const RequireNewPassword: React.FC = () => {
-    const { checkingUser } = useAuthenticator();
     const { shouldRender, requiredAttributes, confirm } = useRequireNewPassword();
 
-    if (!shouldRender || checkingUser) {
+    if (!shouldRender) {
         return null;
     }